+++ conda/platform_windows_amd64.go
@@ -29,7 +29,7 @@ const (
 )
 
 func MicromambaLink() string {
-	return "https://downloads.robocorp.com/micromamba/v0.7.14/windows64/micromamba.exe"
+	return "https://downloads.robocorp.com/micromamba/v0.8.0/windows64/micromamba.exe"
 }
 
 var (